refactor: Extract ComponentLabel
